refactor(noir-proof-batch): rename new batching functionality to original names
