+++ packages/noir-proof-batch/circuits/batch_2_nodes/src/main.nr
