--- packages/noir-proof-batch/src/vks-new.ts
@@ -1,263 +1,0 @@
-// vk for batching 2 leaves circuit
-export const batch2LeavesCircuitVk = [
-    "0x0000000000000000000000000000000000000000000000000000000000200000",
-    "0x0000000000000000000000000000000000000000000000000000000000000011",
-    "0x0000000000000000000000000000000000000000000000000000000000000001",
-    "0x0000000000000000000000000000000000000000000000000000000000000001",
-    "0x0000000000000000000000000000000000000000000000000000000000000001",
-    "0x0000000000000000000000000000000000000000000000000000000000000002",
-    "0x0000000000000000000000000000000000000000000000000000000000000003",
-    "0x0000000000000000000000000000000000000000000000000000000000000004",
-    "0x0000000000000000000000000000000000000000000000000000000000000005",
-    "0x0000000000000000000000000000000000000000000000000000000000000006",
-    "0x0000000000000000000000000000000000000000000000000000000000000007",
-    "0x0000000000000000000000000000000000000000000000000000000000000008",
-    "0x0000000000000000000000000000000000000000000000000000000000000009",
-    "0x000000000000000000000000000000000000000000000000000000000000000a",
-    "0x000000000000000000000000000000000000000000000000000000000000000b",
-    "0x000000000000000000000000000000000000000000000000000000000000000c",
-    "0x000000000000000000000000000000000000000000000000000000000000000d",
-    "0x000000000000000000000000000000000000000000000000000000000000000e",
-    "0x000000000000000000000000000000000000000000000000000000000000000f",
-    "0x0000000000000000000000000000000000000000000000000000000000000010",
-    "0x000000000000000000000000000000a62726fb8292971d55160467231eae35f6",
-    "0x00000000000000000000000000000000002c6defc38aab97d73ae60ca76969d7",
-    "0x000000000000000000000000000000db91aa61d4438fd1bfa73c584b74b89820",
-    "0x000000000000000000000000000000000003c88048e0942926c4d3d69fc24ec6",
-    "0x00000000000000000000000000000016eced9b9650777399ed827d642d0c1857",
-    "0x000000000000000000000000000000000019488f878d4e53b421b3ef4941de52",
-    "0x00000000000000000000000000000096abc8951deda12710a6fd825decd016b9",
-    "0x00000000000000000000000000000000002effe388d52a8bdfbd57a8b0fa8995",
-    "0x0000000000000000000000000000004f5ccce2af5f8044c5697ca624f5612280",
-    "0x00000000000000000000000000000000002ed0fe89db8dcb5c4a0891c7da78b6",
-    "0x000000000000000000000000000000696187fe40defd6b162a0f409730f8b575",
-    "0x0000000000000000000000000000000000014c57902392b220fd3f624d3afa85",
-    "0x000000000000000000000000000000a57657f76a500827cfcfab53a662101484",
-    "0x000000000000000000000000000000000029b8be47622c74603d7b7cbb9db44f",
-    "0x000000000000000000000000000000e442e9b25d8a7eb51b937ea9257c956545",
-    "0x00000000000000000000000000000000001d09760c2442af3291dbc528485a92",
-    "0x0000000000000000000000000000005ca0dbac1286494608d1f8faf88902a37f",
-    "0x00000000000000000000000000000000000d806fe4657fa06f9120b6370bab91",
-    "0x000000000000000000000000000000ceeadcd6ebca36e754df41c1670ef7aba7",
-    "0x00000000000000000000000000000000001ca1aecadbff34c31da98c62295dc9",
-    "0x000000000000000000000000000000f5d4f00333f01f2cffe7a5a8c80af8b43e",
-    "0x00000000000000000000000000000000002d05c543225c09a332a6dc766c5540",
-    "0x000000000000000000000000000000202093041b27b6a9d19a704b20999bfec3",
-    "0x000000000000000000000000000000000029451ecdc0217117697eb263288aec",
-    "0x000000000000000000000000000000e7ed608dc7139a28fa06c24f11778ad095",
-    "0x00000000000000000000000000000000001d333c9acfdda00558a7a8545c7662",
-    "0x0000000000000000000000000000002884561c3ba6333f3e364e4970ce05ced6",
-    "0x000000000000000000000000000000000007aa6d098fec4e36980ea68706d5aa",
-    "0x0000000000000000000000000000004aaf38021b8efb69c6af5f163b22c0288f",
-    "0x0000000000000000000000000000000000146503edef560354ecbb9531c5ab2d",
-    "0x0000000000000000000000000000005e7ef26ef45616c34d60703857e8a19135",
-    "0x0000000000000000000000000000000000002ed1e24f724fb3974936e0e84473",
-    "0x000000000000000000000000000000ffaa18314893c4f8d6e2fe6fac11499c31",
-    "0x00000000000000000000000000000000000a64e3242b7cd2970b97cd76bf2c2d",
-    "0x0000000000000000000000000000000942498cb40f2739c86c1ea0a0decb71f1",
-    "0x00000000000000000000000000000000001bedd315c45902dedf1b7dfe52ecfb",
-    "0x000000000000000000000000000000dc54d800e8854271d12dbe1e1f07d15c80",
-    "0x00000000000000000000000000000000002f9b42429825413dca7d92c776ae74",
-    "0x000000000000000000000000000000b1953b84642851c41ffe9028950f66e3a0",
-    "0x00000000000000000000000000000000002dd7cf27e965dfc7a00397b86d3c74",
-    "0x000000000000000000000000000000d64e5431c3bcf7ce06a29c066c5b483899",
-    "0x000000000000000000000000000000000004b746002c9a6378371781d5546453",
-    "0x000000000000000000000000000000c6d4d7362ca85230e8cd947e3753137c7a",
-    "0x000000000000000000000000000000000028bd3f2d5bea93a3f9e8dd1bd5d956",
-    "0x000000000000000000000000000000d2f1adf73e009ff80a179a150abcfcbbf5",
-    "0x00000000000000000000000000000000002b0b7b206116db428b692658d3e4c2",
-    "0x0000000000000000000000000000006c621dfc6e929c222d3b633d6667cc8371",
-    "0x0000000000000000000000000000000000064ea33275496df61527c83a4f9adc",
-    "0x0000000000000000000000000000009a94180d5360837f067f79ac6670baf71b",
-    "0x000000000000000000000000000000000001c900e51172838bfa0be065d092d0",
-    "0x0000000000000000000000000000002b4a0724903d9c572490737cdfa8d7bd67",
-    "0x0000000000000000000000000000000000115d33697c237f729ad28f54fcb65c",
-    "0x000000000000000000000000000000b29eaf6bdc29ed4e96f62f2df6b63a868e",
-    "0x00000000000000000000000000000000000296e9d71a0b126e591894adbb684e",
-    "0x00000000000000000000000000000031d34fc36bec786bd229b21d7214115a48",
-    "0x00000000000000000000000000000000000291a221d96c29d9c286382b1ebb56",
-    "0x000000000000000000000000000000bbd88e925b355b018f954ef757b3007a12",
-    "0x000000000000000000000000000000000015db0b7ffc6d365761e73484cc7bba",
-    "0x00000000000000000000000000000006461e594806afcee6211bbf50c4a61be0",
-    "0x000000000000000000000000000000000002b87cd06300df41436a9d34e5a7a7",
-    "0x00000000000000000000000000000090b33d8c9070f447ea60b2b044a44af23e",
-    "0x0000000000000000000000000000000000199da44c06272e51502b2080cb4a07",
-    "0x00000000000000000000000000000078442fe7c043e40cfe08af0453df9f9c31",
-    "0x00000000000000000000000000000000002a0302b0e791dfb3c223a6ad42d02f",
-    "0x000000000000000000000000000000b5d8f2a48353d4bb4a9566980437a4759a",
-    "0x000000000000000000000000000000000029ab8fdbd1c76e32a5e1419ac16fe1",
-    "0x00000000000000000000000000000088f1d156bce97e1ffeb0f68ef4c5df828f",
-    "0x00000000000000000000000000000000001440aedf4ef3082bdc4d924764d646",
-    "0x000000000000000000000000000000df11f430e46796871698df834ec7ed0bb3",
-    "0x00000000000000000000000000000000002768a767a05db946514711a349b384",
-    "0x0000000000000000000000000000006a905f76862a44928563465398da26ca38",
-    "0x00000000000000000000000000000000001f3209be8febd91fe2cc050f12b82e",
-    "0x000000000000000000000000000000218521f2993f23bfbcef9b0dc80657393e",
-    "0x00000000000000000000000000000000002289b726027ae662dd742aa691d957",
-    "0x0000000000000000000000000000003d21800e9e38542419001b6c661430b987",
-    "0x0000000000000000000000000000000000116e3ca72a5efaee41c3ef23d1e296",
-    "0x0000000000000000000000000000002ad60f705aa1cf7c03cb9a35d4d6ae1f63",
-    "0x0000000000000000000000000000000000283671fd27b3002a9e193eaf9c7775",
-    "0x000000000000000000000000000000dbda3da393583f0349db84d2083f83e16f",
-    "0x000000000000000000000000000000000001ab7d84d6ce19156b94a2bb287cf3",
-    "0x0000000000000000000000000000004c0fa889c8497180220be3218d1aa877a7",
-    "0x00000000000000000000000000000000001c59c1683eeec78f0957ec81ece0af",
-    "0x000000000000000000000000000000dbf4cfdec4f9af2eaa47ba1b08dc31db3a",
-    "0x00000000000000000000000000000000001e7ef31b81b80af3b54473384e24a8",
-    "0x00000000000000000000000000000004e044728fa53c2fed6c38295c6c074624",
-    "0x00000000000000000000000000000000001de9a7b7e12288822cd9fca620e3b6",
-    "0x000000000000000000000000000000e900dff116039cdb51d3ae038780928bec",
-    "0x000000000000000000000000000000000006dfcf6a6d980d795afcc4a82318ac",
-    "0x000000000000000000000000000000a1a5328ab2f820f5443744220b8932be64",
-    "0x0000000000000000000000000000000000204bb58d4761ee00eecade36351cea",
-    "0x00000000000000000000000000000072163b8a30e9438868b868c6721b3ceb7d",
-    "0x00000000000000000000000000000000000521e2358ee27e15694c703ea302f7",
-    "0x000000000000000000000000000000668803cd8bbdb89cf46f6fc8c8096b6405",
-    "0x00000000000000000000000000000000002bc5ca7c1d1eb3dec80156cd6785d1",
-    "0x0000000000000000000000000000005c21c70fde7a4bfeb7b5ee35faa87c50c9",
-    "0x000000000000000000000000000000000019276638b3e1ec58475359646d7307",
-    "0x0000000000000000000000000000001d7d5f1ab9ff0ed65f753bd39d02c22dd5",
-    "0x00000000000000000000000000000000001d97ed0df61a8150d283002322f8ff",
-    "0x000000000000000000000000000000b151b3ee4ac1fcecbb92c5842d0e3cb220",
-    "0x0000000000000000000000000000000000228199fe8f798f982b4d0ddc1be138",
-    "0x0000000000000000000000000000000000000000000000000000000000000001",
-    "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "0x0000000000000000000000000000000000000000000000000000000000000002",
-    "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "0x0000000000000000000000000000008e24ebe2fae04401bd69c326ca2257c52a",
-    "0x00000000000000000000000000000000002cfb60a802f501040985c638162ef3",
-    "0x0000000000000000000000000000006991473060bd1fd815de60e223ba64c04b",
-    "0x0000000000000000000000000000000000304f0e8a051432fb384b5249b98a59"
-]
-
-// vk for batching 2 nodes circuit
-export const batch2NodesCircuitVk = [
-    "0x0000000000000000000000000000000000000000000000000000000000200000",
-    "0x0000000000000000000000000000000000000000000000000000000000000011",
-    "0x0000000000000000000000000000000000000000000000000000000000000001",
-    "0x0000000000000000000000000000000000000000000000000000000000000001",
-    "0x0000000000000000000000000000000000000000000000000000000000000001",
-    "0x0000000000000000000000000000000000000000000000000000000000000002",
-    "0x0000000000000000000000000000000000000000000000000000000000000003",
-    "0x0000000000000000000000000000000000000000000000000000000000000004",
-    "0x0000000000000000000000000000000000000000000000000000000000000005",
-    "0x0000000000000000000000000000000000000000000000000000000000000006",
-    "0x0000000000000000000000000000000000000000000000000000000000000007",
-    "0x0000000000000000000000000000000000000000000000000000000000000008",
-    "0x0000000000000000000000000000000000000000000000000000000000000009",
-    "0x000000000000000000000000000000000000000000000000000000000000000a",
-    "0x000000000000000000000000000000000000000000000000000000000000000b",
-    "0x000000000000000000000000000000000000000000000000000000000000000c",
-    "0x000000000000000000000000000000000000000000000000000000000000000d",
-    "0x000000000000000000000000000000000000000000000000000000000000000e",
-    "0x000000000000000000000000000000000000000000000000000000000000000f",
-    "0x0000000000000000000000000000000000000000000000000000000000000010",
-    "0x000000000000000000000000000000f656dcb73647b9685d62522e93e92cf7eb",
-    "0x00000000000000000000000000000000001aa9020d2ffe48a5d4e61eb2676b73",
-    "0x000000000000000000000000000000ecd1d7f78f900de62798762645b5593780",
-    "0x000000000000000000000000000000000011157ba99d695274ff68e9008255e4",
-    "0x000000000000000000000000000000549df2c7dcdef37266c0330166ec763f4e",
-    "0x00000000000000000000000000000000000e65b70038ce6dcf8f91d8e781ebed",
-    "0x00000000000000000000000000000012c190c65903d646769491888722c6fd0d",
-    "0x000000000000000000000000000000000024bc4ec179519918bab93327b04a97",
-    "0x000000000000000000000000000000f569912920b597986dab15bdfef70cd8fa",
-    "0x000000000000000000000000000000000000af76b029a085396de14df021bcbe",
-    "0x000000000000000000000000000000c4bdd45dd2af3205713ba9e340dc432db1",
-    "0x00000000000000000000000000000000001efe0d41f24dc916c7ced45e9e8024",
-    "0x0000000000000000000000000000002c131578c45d9f35ee2a9c2dbd0ca284f6",
-    "0x00000000000000000000000000000000002e70c3990bf4d1ef95023e56c6ba10",
-    "0x00000000000000000000000000000097249e4341a579a76127f2f85bd9fb22a3",
-    "0x00000000000000000000000000000000001e64cc6f32994852e5e87ec14e161d",
-    "0x000000000000000000000000000000a00c20ee6721f6c4c0cb0ffd0ba8cfbaae",
-    "0x00000000000000000000000000000000002cfe5b7785d0df03a8330cd53f7b7d",
-    "0x000000000000000000000000000000e982f77f008c4a46094186a9c12d0b9ba4",
-    "0x000000000000000000000000000000000026c4b1a9e7399b92a4afd2b800868a",
-    "0x000000000000000000000000000000895869b0440ac242899def40326dde197d",
-    "0x000000000000000000000000000000000013d2aa6ec85b172fadcca3af8d403b",
-    "0x0000000000000000000000000000002f6866ec217b14e2b80ddee32ff2372002",
-    "0x0000000000000000000000000000000000253e1594dc07e2f1a3cba9a7f6b831",
-    "0x000000000000000000000000000000db452512ad93067821d8e4b1ce0c2e6f23",
-    "0x0000000000000000000000000000000000278ee6c0ec01157d9efabd02fd62d1",
-    "0x0000000000000000000000000000003635601d032ed26f7227aa5a2100b53e4b",
-    "0x00000000000000000000000000000000001e030c621c04be8d3d7c27e86bf383",
-    "0x0000000000000000000000000000005d460dbaaeb77cca4afcb75f81085991ee",
-    "0x000000000000000000000000000000000021362987ae91a50fd2b484ed201bf7",
-    "0x000000000000000000000000000000369aa842f001c2e0b14d3914d70b7da616",
-    "0x0000000000000000000000000000000000023bfd169e77fc8a1341c625680a43",
-    "0x000000000000000000000000000000c0dd89424f3065e0fb258502d7c3149b06",
-    "0x000000000000000000000000000000000029a99131347c75cb45935d2df3f2e2",
-    "0x00000000000000000000000000000023edd963e64cbc870879d7f8a3f0a2a53b",
-    "0x0000000000000000000000000000000000092e06a755aa7f97edd93ca95d1f0c",
-    "0x00000000000000000000000000000034eca6b9f46f7c96e41d99eb2afe059337",
-    "0x000000000000000000000000000000000013d6a0b8a69b0704140760713ef6f2",
-    "0x0000000000000000000000000000009498bd7c36a35d8c295b0755469a2176fb",
-    "0x000000000000000000000000000000000015a4902cbb12c558ca24f2c4e04e5d",
-    "0x000000000000000000000000000000388eb6d0b85a467b16cc4da9f6fb48cd6e",
-    "0x00000000000000000000000000000000001c866d694a291726bbdd01457d2ccb",
-    "0x0000000000000000000000000000001519d0bdca8e99f793b618e9950c888358",
-    "0x0000000000000000000000000000000000280302b4de7028562030662083fc62",
-    "0x000000000000000000000000000000946d660bad976163ac7992ec7230d6ee3f",
-    "0x00000000000000000000000000000000000a150da86013d37b90e7de36496f44",
-    "0x00000000000000000000000000000026833c6b7b0fc21ab34001daf68069419c",
-    "0x00000000000000000000000000000000001593f7e8ef9111dc93b91f19022b6d",
-    "0x000000000000000000000000000000e4f869e089f1ff403d2932e6c9f72b201c",
-    "0x00000000000000000000000000000000001ff7fd70f4c2fb5d17d59ccc80acee",
-    "0x000000000000000000000000000000f99ee822e13f6d92ea76512befcba4cb36",
-    "0x00000000000000000000000000000000002f20fca8be2502ddfa648ef137b5a4",
-    "0x000000000000000000000000000000ede079ac66bd5e0a3b6365a36d41d13001",
-    "0x000000000000000000000000000000000021133e2f7fbf559fcabed807422c8d",
-    "0x000000000000000000000000000000424724c586d1843426f49f8a9899bea110",
-    "0x000000000000000000000000000000000029625c645c01360da28f8edbdae828",
-    "0x000000000000000000000000000000912c114b42f7d7a6459d46e940574707f0",
-    "0x000000000000000000000000000000000022ae2116f08eee3a15dfa2cc3de5c3",
-    "0x000000000000000000000000000000cf456eec08bbf0941b7a8dad10c3948ecb",
-    "0x000000000000000000000000000000000001c2b204d2026e697ae79cbebf9541",
-    "0x0000000000000000000000000000003be8ea8360a3a45e8303231f1069f4f524",
-    "0x00000000000000000000000000000000000aaa5912819926b550505b37f4e0fa",
-    "0x0000000000000000000000000000007713dccb2f2d1f9cd9dd5ff6260c010fb5",
-    "0x000000000000000000000000000000000028fb656e68f4d14622cefd36cfd57b",
-    "0x0000000000000000000000000000001430111697ff777a045182da5a783b52be",
-    "0x0000000000000000000000000000000000199b36bd23bda2b389049319fd80d0",
-    "0x000000000000000000000000000000006960110374616f4aa0257a64626b91ee",
-    "0x0000000000000000000000000000000000058ea6af2bae2b2ad05974a50b0826",
-    "0x000000000000000000000000000000772db7f2e0f46f3a31f630cdacdf29fe6a",
-    "0x00000000000000000000000000000000000f14040ac3a31a4ab1a209d059d91b",
-    "0x0000000000000000000000000000002039e7f0e5c7acea108f181b9c60c137d7",
-    "0x00000000000000000000000000000000000ec65e9f07d9e5066b5a14ce9259cc",
-    "0x000000000000000000000000000000d6064d1e4c5657f2c9178c538b5b6fcac0",
-    "0x000000000000000000000000000000000022713f12fac4974e9372d96b0df6c4",
-    "0x000000000000000000000000000000e16d463cabdf9006152137f365a9c0bef5",
-    "0x00000000000000000000000000000000000c0e2b246768e44607e732b8f69ce6",
-    "0x000000000000000000000000000000b4109059dd670060a88fbf3ea619a865f6",
-    "0x0000000000000000000000000000000000172006483fadd4bba36a75950ec726",
-    "0x000000000000000000000000000000d300851943f1b2db78e9812135b9c65e27",
-    "0x00000000000000000000000000000000002b2ae015863d122ac04b0947f6f565",
-    "0x000000000000000000000000000000198aadd3c70cbdc20f9ea42456047cd12e",
-    "0x000000000000000000000000000000000008dd4bc155e6e8c6b0caf22631aad4",
-    "0x000000000000000000000000000000c801731d432e0f32e8ffa05116b7729e60",
-    "0x00000000000000000000000000000000000d20e193ba18173d9b939fe59cf669",
-    "0x000000000000000000000000000000e9b586c7a328d999429ed5121cb66a044e",
-    "0x000000000000000000000000000000000006ad602dbefb48e4fd97ad3e6b26cf",
-    "0x0000000000000000000000000000007ed19e80fc500fa06e2ee82b3da63999df",
-    "0x00000000000000000000000000000000002d6e23f9c535c3c78030ca87620c5b",
-    "0x0000000000000000000000000000004666d9a925170fb5b12c7b23808b161479",
-    "0x00000000000000000000000000000000002c71e0ca8ce932bbc1bc27cc01d3bf",
-    "0x00000000000000000000000000000008c9b837b26d4bbadd003509ba019b8d8c",
-    "0x0000000000000000000000000000000000113db4e7a30a943282c2a1e8a5e63b",
-    "0x0000000000000000000000000000009be77690fe4bbf969411f04554b5297097",
-    "0x000000000000000000000000000000000016f4fd08e3c1b1ef24a475831dac8b",
-    "0x000000000000000000000000000000e872959e21a0af5377cfc9ae7266c8b1be",
-    "0x000000000000000000000000000000000010e66bf72fd0b9a0c436caf8306d03",
-    "0x0000000000000000000000000000003fe7768092818a06cdf46f2c6ce09fe79e",
-    "0x00000000000000000000000000000000001f493f573098c071d2f5b58db4e8b9",
-    "0x000000000000000000000000000000f1711f000d503c918911d343911d04d32c",
-    "0x000000000000000000000000000000000015796263418edb897ecd6891ca61c3",
-    "0x0000000000000000000000000000000000000000000000000000000000000001",
-    "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "0x0000000000000000000000000000000000000000000000000000000000000002",
-    "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "0x0000000000000000000000000000004d2c76d44b6f49e3638103874d37300a0e",
-    "0x000000000000000000000000000000000025fc4c7bf131eb404f5cb90c57933b",
-    "0x000000000000000000000000000000d30cbb1e83c2e804d09f19ad4cca8c14e9",
-    "0x00000000000000000000000000000000002a25fce58d87d78ee12d91bdb3fb7f"
-]
